Fixing types, first round.
Some internals headers and ap/plan9 sources

Removed plan9/brk.c now brk/sbrk come from libc (tested ok)
Added chdir syscall wrapper, chdir() ansi func is based on it!

Updated posix version in unistd.h

+++ amd64/include/ureg.h
@@ -5,21 +5,21 @@
 #endif
 
 struct Ureg {
-	unsigned long long	ax;
-	unsigned long long	bx;
-	unsigned long long	cx;
-	unsigned long long	dx;
-	unsigned long long	si;
-	unsigned long long	di;
-	unsigned long long	bp;
-	unsigned long long	r8;
-	unsigned long long	r9;
-	unsigned long long	r10;
-	unsigned long long	r11;
-	unsigned long long	r12;
-	unsigned long long	r13;
-	unsigned long long	r14;
-	unsigned long long	r15;
+	uint64_t	ax;
+	uint64_t	bx;
+	uint64_t	cx;
+	uint64_t	dx;
+	uint64_t	si;
+	uint64_t	di;
+	uint64_t	bp;
+	uint64_t	r8;
+	uint64_t	r9;
+	uint64_t	r10;
+	uint64_t	r11;
+	uint64_t	r12;
+	uint64_t	r13;
+	uint64_t	r14;
+	uint64_t	r15;
 
 	// these are pointless and I don't want to waste a second on them.
 	// I'm sure it will break something but let's see.
@@ -29,13 +29,13 @@ struct Ureg {
 	//	unsigned short	fs;
 	//	unsigned short	gs;
 
-	unsigned long long	type;
-	unsigned long long	error;		/* error code (or zero) */
-	unsigned long long	ip;		/* pc */
-	unsigned long long	cs;		/* old context */
-	unsigned long long	flags;		/* old flags */
-	unsigned long long	sp;		/* sp */
-	unsigned long long	ss;		/* old stack segment */
+	uint64_t	type;
+	uint64_t	error;		/* error code (or zero) */
+	uint64_t	ip;		/* pc */
+	uint64_t	cs;		/* old context */
+	uint64_t	flags;		/* old flags */
+	uint64_t	sp;		/* sp */
+	uint64_t	ss;		/* old stack segment */
 };
 
 #endif
